+++ openrouteservice/src/main/java/heigit/ors/routing/RoutingProfileManager.java
@@ -30,9 +30,6 @@ import heigit.ors.mapmatching.MapMatchingRequest;
 import heigit.ors.matrix.MatrixErrorCodes;
 import heigit.ors.matrix.MatrixRequest;
 import heigit.ors.matrix.MatrixResult;
-import heigit.ors.optimization.OptimizationErrorCodes;
-import heigit.ors.optimization.RouteOptimizationRequest;
-import heigit.ors.optimization.RouteOptimizationResult;
 import heigit.ors.routing.configuration.RouteProfileConfiguration;
 import heigit.ors.routing.configuration.RoutingManagerConfiguration;
 import heigit.ors.routing.pathprocessors.ElevationSmoothPathProcessor;
@@ -609,12 +606,4 @@ public class RoutingProfileManager {
         return rp.computeMatrix(req);
     }
 
-    public RouteOptimizationResult computeOptimizedRoutes(RouteOptimizationRequest req) throws Exception {
-        RoutingProfile rp = _routeProfiles.getRouteProfile(req.getProfileType(), true);
-
-        if (rp == null)
-            throw new InternalServerException(OptimizationErrorCodes.UNKNOWN, "Unable to find an appropriate routing profile.");
-
-        return rp.computeOptimizedRoutes(req);
-    }
 }