+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Updated pom to always build ors.war (Issue #432)
 - Replace usage of packages incompatible with Java >8 (#474)
 - Moved walking and hiking flag encoders to the ORS core system (#440)
+- Remove route optimization code (#499)
 ### Deprecated